fix CR fixes and refactor code
